Fixed bug for Issue #50.
By setting the provider on the server worked for Ubuntu 12.04, 14.04,
and 15.04.  I was unable to get 10.04 working (there were some problems
outside the scope of this issue with the code that 10.04 has).

However, I ran into a new issue with 15.04.  The following bug report
best describes the problem:

https://bugs.launchpad.net/ubuntu/+source/nfs-utils/+bug/1428961

Essentially, idmapd no longer as an init script and therefore fails. The
nfs-server init script manages idmapd so I used a case statement to
remove this functionality from Ubuntu 15.04 and later versions.

Finially, I added a vagrant directory that contains the testing
framework that I used to manually test all of the supported version
(10.04, 12.04, 14.04, and 15.04).  I had to modify the tests/init.pp to
test both the server and client.